minor performance changes
Should not have to check vaulted list if the user has not selected to show that information. Offset should start at 140 pixels to save on loops as thats the minimum size supported in the first place. Will be making a debug version at some point that draws on the screen